requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.0
yfinance==0.2.31
beautifulsoup4==4.12.2
requests==2.31.0
langchain==0.0.350
python-dotenv==1.0.0